requirements.txt
Django>=4.2
djangorestframework>=3.15
psycopg2-binary>=2.9
python-dotenv>=1.0
